Improve predicate logic for `in` operator
This improves the performance when the Parquet files' statistics
are between the values provided to "in" but none of the values provided
to "in" are between the Parquet statistics.

Current implementation is as follows:
e.g.
`predicates = [[("x", "in", (13, 15, 20))]]`
Given a Parquet file `A.parquet` with statistics: x.min = 17, x.max=18

The predicate is parsed as x >= 13, x <= 20 in predicate logic. This would
yield true since according to Parquet the range of x goes from 17 to 18.

`A.parquet` would be read entirely, but that is unnecessary given the predicates.

This change parses the provided predicate as:
`any([x.min <= v <= x.max for v in (13, 15, 20)])`
Thus, `A.parquet` would not need to be read entirely, as there is no
chance for values of `A` to be in the range specified