Disable pip interop in CI
Fixes #381

Otherwise `conda` thinks it needs to upgrade the installed `kartothek` version. We otherwise only use conda packages in the CI and thus the best way is to tell conda not to touch any `pip`/`pypi` installed packages instead of guessing what to do with them. The problems in #380 didn't occur much earlier as `conda` always installed the latest release and that was mostly fine in CI.