Add optional Parquet serializer to cube interface
This allows to specify the compression algorithm and a chunk size to
generate row groups.

The Parquet serializer parameter is added the following cube operations:
  * `build_cube`
  * `append_cube`
  * `extend_cube`
  * `update_cube`

Specifying different Parquet serializers is back- and forwards compatible.
Row groups are kept for partitions that are not touched during cube operations.
If partitions are replaced or newly created, they are rewritten by the
specified Parquet serializer.